+++ app/api/core/infrastructure/mongodb/template/Mapper.ts
